refactor(handlers): align JumpShip imports with WarShip topology
Extract IJumpShip into a new sibling leaf module
JumpShipUnitHandler.types.ts so the .calculations.ts and .helpers.ts
leaves no longer back-import the orchestrator. This mirrors WarShip,
where the type lives outside the orchestrator and the leaves are
pure (never importing from .ts).

Closes 2 of the 4 circular dependencies in
src/services/units/handlers/:
  - JumpShipUnitHandler.calculations.ts <-> JumpShipUnitHandler.ts
  - JumpShipUnitHandler.helpers.ts <-> JumpShipUnitHandler.ts

Public API preserved: IJumpShip is re-exported from
JumpShipUnitHandler.ts (and through the index.ts barrel) so existing
consumers compile unchanged. BV parity verified via
scripts/validate-bv.ts: per-unit results byte-identical
pre/post-refactor.
